pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = "Portfolio Quantitative Finance"
authors = ["Javier Campoy <jcampoyvilchez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
fastapi = {extras = ["all"], version = "^0.108.0"}
sqlalchemy = "^2.0.24"
numpy = "^1.26.3"
mysql-connector-python = "^8.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pylint = "^3.0.2"
black = "^23.11.0"
isort = "^5.13.0"
mypy = "^1.7.1"
pytest-cov = "^4.1.0"
ipdb = "^0.13.13"
pandas-stubs = "^2.1.1.230928"
types-requests = "^2.31.0.10"
types-python-dateutil = "^2.8.19.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

#[tool.mypy]
#ignore_missing_imports = false